requirements.txt
django==2.0
django-hosts==3.0
django-crispy-forms==1.7.2
django-import-export==1.0.1
openpyxl==2.4.9
odfpy==1.3.6
jdcal==1.4
xlwt==1.3.0
pyyaml==3.13
python-decouple==3.1
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
djangorestframework-datatables==0.4.0
djangorestframework==3.8.2
